% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{binomial_matrix}
\alias{binomial_matrix}
\title{Create binomial matrix}
\usage{
binomial_matrix(n)
}
\arguments{
\item{n}{- row dimension}
}
\description{
Binomial matrix: an N-by-N multiple of an involuntory matrix with
integer entries such that $A^2 = 2^(N-1)*I_N$
Thus B = A*2^((1-N)/2) is involutory, that is B^2 = EYE(N)
}
