% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{cauchy_matrix}
\alias{cauchy_matrix}
\title{Create cauchy matrix}
\usage{
cauchy_matrix(x, y = NULL)
}
\arguments{
\item{x}{vector of length n}

\item{y}{vector of length n}
}
\description{
Arguments \code{x} and \code{y} are vectors of length \code{n}. \code{C[i,j] = 1 / (x[i] + y[j])}
}
