% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{chow}
\alias{chow}
\title{Creating singular Toeplitz lower Hessenberg matrix}
\usage{
chow(n, alpha = 1, delta = 0)
}
\arguments{
\item{n}{order of the matrix}

\item{alpha}{defaults to 1}

\item{delta}{defaults to 0}
}
\description{
returns matrix \code{A = H(alpha) + delta * EYE}, such that
  \code{H[i,j] = alpha^(i-j+1)}.
}
