% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{circul}
\alias{circul}
\title{Create circulant matrix}
\usage{
circul(v)
}
\arguments{
\item{v}{first row of the matrix. If \code{v} is a scalar, then \code{C = circul(1:v)}}
}
\value{
a circulant matrix whose first row is the vector \code{v}
}
\description{
Each row is obtained from the previous by cyclically permuting the
  entries one step forward. A special Toeplitz matrix in which diagonals "wrap around"
}
