% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{clement}
\alias{clement}
\title{Create Clement tridiagonal matrix with zero diagonal entries}
\usage{
clement(n, k = 0)
}
\arguments{
\item{n}{order of matrix}

\item{k}{0 indicates symmetric matrix, 1 asymmetric}
}
\description{
Returns an \code{n}-by-\code{n} tridiagonal matrix with zeros on the
  main diagonal. For \code{k=0}, \code{A} is nonsymmetric. For \code{k=1, A} is
  symmetric
}
