% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{compar}
\alias{compar}
\title{Create comparison matrix \code{A}}
\usage{
compar(B, k = 0)
}
\arguments{
\item{B}{input matrix}

\item{k}{decides what matrix to return}
}
\description{
For \code{k=0}, if \code{i==j}, $A[i,j]=abs(B[i,j])$ and
  \code{A[i,j]=-abs(B[i,j])} otherwise. For \code{k=1}, \code{A} replaces each
  diagonal element of \code{B} with its absolute value, and replaces each
  off-diagonal with the negative of the largest absolute value off-diagonal
  in the same row.
}
