% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{cycol}
\alias{cycol}
\title{Create matrix \code{A} whose columns repeat cyclically}
\usage{
cycol(n, k, m = NULL)
}
\arguments{
\item{n}{number of columns of matrix}

\item{k}{upper limit of rank}

\item{m}{number of rows of matrix}
}
\description{
Returns an \code{n}-by-\code{n} matrix with cyclically repeating columns
  where one cycle consists of the columns defined by \code{randn(n,k)}. Thus, the
  rank of matrix \code{A} cannot exceed \code{k}, and \code{k} must be scalar.
}
