% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{dorr}
\alias{dorr}
\title{Create Dorr matrix}
\usage{
dorr(n, theta = 0.01)
}
\arguments{
\item{n}{order of matrix}

\item{theta}{determines conditionality. Ill-conditioned when theta is nonnegative.}
}
\description{
Returns a \code{n}-by-\code{n} row diagonally dominant,
  tridiagonal matrix that is ill-conditioned for small nonnegative values
  of \code{theta}. The default value of \code{theta} is 0.01.
}
