% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{fiedler}
\alias{fiedler}
\title{Create Fiedler matrix}
\usage{
fiedler(c)
}
\arguments{
\item{c}{N-vector. If \code{c} is a scalar, then returns fiedler(1:c)}
}
\value{
a symmetric dense matrix A with a dominant positive eigenvalue and all others are negative.
}
\description{
Fiedler matrix that has a dominant positive eigenvalue and all others are negative
}
