% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{forsythe}
\alias{forsythe}
\title{Create Forsythe matrix or perturbed Jordan block}
\usage{
forsythe(n, alpha = .Machine$double.eps, lambda = 0)
}
\arguments{
\item{n}{order of matrix}

\item{alpha}{value of perturbation at \code{A[n,1]}}

\item{lambda}{eigenvalue of Jordan block}
}
\description{
Returns a \code{n}-by-\code{n} matrix equal to the Jordan block with
  eigenvalue \code{lambda}, except that \code{A[n,1]=alpha}.
}
