% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_to_pdf}
\alias{ggplot_to_pdf}
\title{Printing ggplot objects to PDF's}
\usage{
ggplot_to_pdf(plot, ...)
}
\arguments{
\item{plot}{a plot object. ggplot most common}

\item{...}{Arguments to pass to \code{pdf}}
}
\description{
Printing plots to a pdf
}
