% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{leslie}
\alias{leslie}
\title{Create Leslie population model matrix}
\usage{
leslie(a, b = NULL, sparse = F)
}
\arguments{
\item{a}{average birth numbers (first row)}

\item{b}{survival rates (subdiagonal)}

\item{sparse}{whether to return a sparse matrix}
}
\value{
N by N Leslie population model matrix
}
\description{
N by N matrix from Leslie population model with average birth and survival rates.
}
