% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{linScale}
\alias{linScale}
\title{Linear scaling of numeric vectors}
\usage{
linScale(x, newmin = 0, newmax = 1)
}
\arguments{
\item{x}{numeric vector}

\item{newmin}{new minimum}

\item{newmax}{new maximum}
}
\value{
A numeric vector linearly scaled of the input to be between \code{newmin} and \code{newmax}.
}
\description{
Linear transformation of a numeric vector to a
newly set minimum and maximum.
}
\examples{
# Scale samples from standard normal between 0 and 1
x <- rnorm(100)
scaled_x <- linScale(x, 0, 1)
plot(x, scaled_x)
}
