% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{plotPseudospectra}
\alias{plotPseudospectra}
\title{Visualize pseudospectra of a matrix}
\usage{
plotPseudospectra(
  A,
  title = "e-pseudospectrum",
  eps = c(1, 0.5, 0.1, 0.01, 0.001)
)
}
\arguments{
\item{A}{a square matrix}

\item{title}{title of the pseudospectra plot}

\item{eps}{epsilon values}
}
\value{
a ggplot object
}
\description{
Plots contour plot on the complex plane for the
pseudospectra for the matrix A, under the given epsilon.
}
