% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line.R
\name{setwd_jump}
\alias{setwd_jump}
\title{setwd_jump}
\usage{
setwd_jump(dir)
}
\arguments{
\item{dir}{-- name of directory up the chain of the current working directory
to change to}
}
\value{
None
}
\description{
Changes working directory to folder up the chain, without
needing to type full filepath or series of "../../.."
}
