% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{simNew.glm}
\alias{simNew.glm}
\title{Generate simulations of response of a fitted \code{glm} object on new data}
\usage{
\method{simNew}{glm}(object, ...)
}
\arguments{
\item{object}{object of class inheriting from \code{glm}}

\item{...}{further arguments passed to or from other methods}
}
\value{
a \code{nrow(newdata) x nsim} sized matrix. Each row corresponds to an observation in the new dataset,
  and each colulmn corresponds to a draw/simulation.
}
\description{
Returns a matrix of simulated responses from a fitted \code{glm} object, applied to a new dataset
  separate from the original data used to fit the model.
}
