% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{simNew}
\alias{simNew}
\title{Generate simulations of response of a fitted model object on new data}
\usage{
simNew(object, ...)
}
\arguments{
\item{object}{model object for which draws of prediction is desired}

\item{...}{additional arguments affecting the prediction}
}
\description{
\code{simulateDraws} is a generic function for generating simulation predictions
from a model object and new data.
}
