% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{simulateDraws}
\alias{simulateDraws}
\title{Generate simulations of a fitted model object on new data}
\usage{
simulateDraws(object, ...)
}
\arguments{
\item{object}{model object for which draws of prediction is desired}

\item{...}{additional arguments affecting the prediction}
}
\description{
\code{simulateDraws} is a generic function for generating simulation predictions
from a model object and newdata.
}
