% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{spdiags}
\alias{spdiags}
\title{Create sparse diagonal matrix}
\usage{
spdiags(A, d, m, n)
}
\arguments{
\item{A}{matrix where columns correspond to the desired diagonals}

\item{d}{indices of the diagonals to be filled in. 0 is main diagonal. -1
is first subdiagonal and +1 is first superdiagonal.}

\item{m}{row dim}

\item{n}{col dim}
}
\value{
dgcMatrix sparse diagonal
}
\description{
Creates a sparse representation of multiple diagonal matrix
}
