% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{tridiag}
\alias{tridiag}
\title{Create sparse tridiagonal matrix}
\usage{
tridiag(n, x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{n}{dimension of the square matrix}

\item{x}{subdiagonal (-1)}

\item{y}{diagonal (0)}

\item{z}{superdiagonal (+1)}
}
\value{
Sparse tridiagonal matrix
}
\description{
Create a sparse tridiagonal matrix of dgcMatrix class.
}
