% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line.R
\name{winPath}
\alias{winPath}
\title{Change Windows filepaths to R format}
\usage{
winPath(path = "clipboard")
}
\arguments{
\item{path}{Filepath with \\ slashes that needs to be formatted. If left blank, will pull from the clipboard.}
}
\value{
Filepath with forward slashes
}
\description{
Convert Windows backslash filepaths to forward slashes to
be compatible with R
}
\examples{
# If Windows path on my clipboard
winPath()
# Input path
winPath("C:\\\\Users\\\\user")

}
