% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{grcar}
\alias{grcar}
\title{Create Toeplitz matrix with sensitive eigenvalues}
\usage{
grcar(n, k = NULL)
}
\value{
n by n Toeplitz matrix with -1 on subdiagonal, 1 on diagonal, and k superdiagionals of 1s.
}
\description{
Eigenvalues are sensitive.
}
